+++ sandpack-react/src/index.ts
@@ -18,6 +18,7 @@ export * from "./contexts/sandpackContext";
 export * from "./contexts/themeContext";
 export * from "./themes";
 export * from "./types";
-export { getCssText as getSandpackCssText } from "./styles";
+export * from "./icons";
 
+export { getCssText as getSandpackCssText } from "./styles";
 export * from "@code-hike/classer";