+++ sandpack-react/src/types.ts
@@ -1,12 +1,15 @@
 import type {
   BundlerState,
   ListenerFunction,
+  ReactDevToolsMode,
   SandpackBundlerFiles,
   SandpackClient,
   SandpackError,
   SandpackMessage,
   UnsubscribeFunction,
 } from "@codesandbox/sandpack-client";
+
+import type { CodeEditorProps } from ".";
 
 export type SandpackClientDispatch = (
   msg: SandpackMessage,
@@ -47,7 +50,7 @@ export interface SandpackState {
   setActiveFile: (path: string) => void;
   resetFile: (path: string) => void;
   resetAllFiles: () => void;
-  registerReactDevTools: () => void;
+  registerReactDevTools: (value: ReactDevToolsMode) => void;
 
   // Element refs
   // Different components inside the SandpackProvider might register certain elements of interest for sandpack
@@ -83,6 +86,7 @@ export interface SandpackFile {
   code: string;
   hidden?: boolean;
   active?: boolean;
+  readOnly?: boolean;
 }
 
 export type SandpackFiles = Record<string, string | SandpackFile>;
@@ -227,6 +231,22 @@ export interface SandpackTheme {
 export type SandpackPartialTheme = DeepPartial<SandpackTheme>;
 
 export type SandpackThemeProp = SandpackPredefinedTheme | SandpackPartialTheme;
+
+/**
+ * Custom properties to be used in the SandpackCodeEditor component,
+ * some of which are exclusive to customize the CodeMirror instance.
+ */
+export interface SandpackCodeOptions {
+  /**
+   * CodeMirror extensions for the editor state, which can
+   * provide extra features and functionalities to the editor component.
+   */
+  extensions?: CodeEditorProps["extensions"];
+  /**
+   * Property to register CodeMirror extension keymap.
+   */
+  extensionsKeymap?: CodeEditorProps["extensionsKeymap"];
+}
 
 /**
  * @hidden